+++ oap-server/server-storage-plugin/storage-elasticsearch-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/elasticsearch/query/UIMenuManagementEsDAO.java
@@ -18,50 +18,26 @@
 
 package org.apache.skywalking.oap.server.storage.plugin.elasticsearch.query;
 
-import org.apache.skywalking.library.elasticsearch.requests.search.Query;
-import org.apache.skywalking.library.elasticsearch.requests.search.Search;
-import org.apache.skywalking.library.elasticsearch.requests.search.SearchBuilder;
-import org.apache.skywalking.library.elasticsearch.response.search.SearchHit;
-import org.apache.skywalking.library.elasticsearch.response.search.SearchResponse;
 import org.apache.skywalking.oap.server.core.management.ui.menu.UIMenu;
 import org.apache.skywalking.oap.server.core.storage.management.UIMenuManagementDAO;
+import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
 import org.apache.skywalking.oap.server.library.client.elasticsearch.ElasticSearchClient;
-import org.apache.skywalking.oap.server.storage.plugin.elasticsearch.base.ElasticSearchConverter;
-import org.apache.skywalking.oap.server.storage.plugin.elasticsearch.base.EsDAO;
-import org.apache.skywalking.oap.server.storage.plugin.elasticsearch.base.IndexController;
 
 import java.io.IOException;
+import org.apache.skywalking.oap.server.storage.plugin.elasticsearch.base.ManagementCRUDEsDAO;
 
-public class UIMenuManagementEsDAO extends EsDAO implements UIMenuManagementDAO {
-    public UIMenuManagementEsDAO(ElasticSearchClient client) {
-        super(client);
+public class UIMenuManagementEsDAO extends ManagementCRUDEsDAO implements UIMenuManagementDAO {
+    public UIMenuManagementEsDAO(ElasticSearchClient client, StorageBuilder storageBuilder) {
+        super(client, storageBuilder);
     }
 
     @Override
     public UIMenu getMenu(String id) throws IOException {
-        final String index =
-            IndexController.LogicIndicesRegister.getPhysicalTableName(UIMenu.INDEX_NAME);
-        final SearchBuilder search =
-            Search.builder().query(Query.ids(id)).size(1);
-        final SearchResponse response = getClient().search(index, search.build());
-
-        if (response.getHits().getHits().size() > 0) {
-            UIMenu.Builder builder = new UIMenu.Builder();
-            SearchHit data = response.getHits().getHits().get(0);
-            return builder.storage2Entity(new ElasticSearchConverter.ToEntity(UIMenu.INDEX_NAME, data.getSource()));
-        }
-        return null;
+        return (UIMenu) super.getById(UIMenu.INDEX_NAME, id);
     }
 
     @Override
     public void saveMenu(UIMenu menu) throws IOException {
-        try {
-            final UIMenu.Builder builder = new UIMenu.Builder();
-            final ElasticSearchConverter.ToStorage toStorage = new ElasticSearchConverter.ToStorage(UIMenu.INDEX_NAME);
-            builder.entity2Storage(menu, toStorage);
-            getClient().forceInsert(UIMenu.INDEX_NAME, menu.id().build(), toStorage.obtain());
-        } catch (Exception e) {
-            throw new IOException(e.getMessage(), e);
-        }
+        super.create(UIMenu.INDEX_NAME, menu);
     }
 }