+++ docs/en/changes/changes.md
@@ -8,6 +8,8 @@
 * ElasticSearchClient: Add `deleteById` API.
 * Fix Custom alarm rules are overwritten by 'resource/alarm-settings.yml'
 * Support Kafka Monitoring.
+* [Breaking Change] Elasticsearch storage merge all management data indices into one index `management`, 
+  including `ui_template，ui_menu，continuous_profiling_policy`.
 
 #### UI